JSON Writer + Unit Tests

+++ Engine/Source/Core/Public/Misc/CoreLiterals.h
@@ -11,7 +11,7 @@ constexpr SIZE_T operator"" _MB(SIZE_T Value)
     return Value * (1024 * 1024);
 }
 
-constexpr u64 operator"" _GB(u64 Value)
+constexpr SIZE_T operator"" _GB(u64 Value)
 {
     return Value * (1024 * 1024 * 1024);
 }
